time pour l'evenement n'est plus mis à zéro à chaque lancement, on prend, le nombre  de milliseconds de la date. changement du champs dans las base vers bigint

+++ snap/models.py
@@ -58,7 +58,7 @@ class Evenement(models.Model):
     session_key=models.CharField(max_length=40,null=True)
     programme=models.ForeignKey(ProgrammeBase,null=True, on_delete=models.SET_NULL) #programme de base chargé
     type=models.CharField(max_length=3,choices=TYPE_EVENEMENT_CHOICES, default=AUTRE) #type d'évènement
-    time=models.IntegerField() #Temps (local à Snap) de l'évènement
+    time=models.BigIntegerField() #Temps (local à Snap) de l'évènement
     numero=models.IntegerField() #numero d'ordre de l'évènement, indépendant du type
     creation=models.DateTimeField(auto_now_add=True)
     def toD3(self):
